DistributionFits

Documentation for DistributionFits.

DistributionFits.AbstractDistributionFitOptimizer

DistributionFits.AbstractMoments

DistributionFits.AbstractΣstar

DistributionFits.QuantilePoint

DistributionFits.fit_mean_quantile

DistributionFits.fit_mean_relerror

DistributionFits.fit_mean_Σ

DistributionFits.fit_median_quantile

DistributionFits.fit_mode_flat

DistributionFits.fit_mode_quantile

DistributionFits.meanFunOfProb

DistributionFits.moments

DistributionFits.ofLogitNormalModeUpper

DistributionFits.shifloNormal

DistributionFits.σstar

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit

StatsAPI.fit

[comment]: # (```@autodocs)

[comment]: # (Modules = [DistributionFits])

[comment]: # (```)

taken care of in index.md called only internally from fit, documented in docstring via ","

DistributionFits.fit_median_quantile — Function

fit(D, val, qp, ::Val{stats} = Val(:mean))

Fit a statistical distribution to a quantile and given statistics

Arguments

D: The type of distribution to fit

val: The value of statistics

qp: QuantilePoint(q,p)

stats Which statistics to fit: defaults to Val(:mean). Alternatives are: Val(:mode), Val(:median)

Examples

d = fit(LogNormal, 5.0, @qp_uu(14));
(mean(d),quantile(d, 0.975)) .≈ (5,14)

d = fit(LogNormal, 5.0, @qp_uu(14), Val(:mode));
(mode(d),quantile(d, 0.975)) .≈ (5,14)

DistributionFits.fit_mean_quantile — Function

fit(D, val, qp, ::Val{stats} = Val(:mean))

Fit a statistical distribution to a quantile and given statistics

Arguments

D: The type of distribution to fit

val: The value of statistics

qp: QuantilePoint(q,p)

stats Which statistics to fit: defaults to Val(:mean). Alternatives are: Val(:mode), Val(:median)

Examples

d = fit(LogNormal, 5.0, @qp_uu(14));
(mean(d),quantile(d, 0.975)) .≈ (5,14)

d = fit(LogNormal, 5.0, @qp_uu(14), Val(:mode));
(mode(d),quantile(d, 0.975)) .≈ (5,14)

DistributionFits.fit_mode_quantile — Function

fit(D, val, qp, ::Val{stats} = Val(:mean))

Fit a statistical distribution to a quantile and given statistics

Arguments

D: The type of distribution to fit

val: The value of statistics

qp: QuantilePoint(q,p)

stats Which statistics to fit: defaults to Val(:mean). Alternatives are: Val(:mode), Val(:median)

Examples

d = fit(LogNormal, 5.0, @qp_uu(14));
(mean(d),quantile(d, 0.975)) .≈ (5,14)

d = fit(LogNormal, 5.0, @qp_uu(14), Val(:mode));
(mode(d),quantile(d, 0.975)) .≈ (5,14)

internals

DistributionFits.meanFunOfProb — Function

compute mean over a function(uniformly distributed probabilities)

used to estimate moments of logitnorm

DistributionFits.ofLogitNormalModeUpper — Function

objective function used by matchModeUpper(LogitNormal,...)

[comment]: # (fit(Type{LogNormal}, Any, AbstractΣstar))