DistributionFits

Documentation for DistributionFits.

DistributionFits.AbstractMoments

DistributionFits.AbstractΣstar

DistributionFits.fit_mean_quantile

DistributionFits.fit_median_quantile

DistributionFits.fit_mode_flat

DistributionFits.fit_mode_flat

DistributionFits.meanFunOfProb

DistributionFits.moments

DistributionFits.ofLogitNormalModeUpper

DistributionFits.σstar

StatsBase.fit

StatsBase.fit

StatsBase.fit

StatsBase.fit

StatsBase.fit

DistributionFits.fit_mean_quantile — Method

fit(D, val, qp, ::Val{stats} = Val(:mean))

Fit a statistical distribution to a quantile and given statistics

Arguments

D: The type of distribution to fit

val: The value of statistics

qp: QuantilePoint(q,p)

stats Which statistics to fit: defaults to Val(:mean). Alternatives are: Val(:mode), Val(:median)

Examples

d = fit(LogNormal, 5, @qp_uu(14));
(mean(d),quantile(d, 0.975)) .≈ (5,14)

d = fit(LogNormal, 5, @qp_uu(14), Val(:mode));
(mode(d),quantile(d, 0.975)) .≈ (5,14)

DistributionFits.fit_median_quantile — Method

fit(D, lower, upper)

Fit a statistical distribution to a set of quantiles

Arguments

D: The type of the distribution to fit

lower: lower QuantilePoint (p,q)

upper: upper QuantilePoint (p,q)

Notes

Several macros help to construct QuantilePoints

@qp(q,p) quantile at specified p: QuantilePoint(q,p)

@qp_ll(q0_025) quantile at very low p: QuantilePoint(q0_025,0.025)

@qp_l(q0_05) quantile at low p: QuantilePoint(q0_05,0.05)

@qp_m(median) quantile at median: QuantilePoint(median,0.5)

@qp_u(q0_95) quantile at high p: QuantilePoint(q0_95,0.95)

@qp_uu(q0_975) quantile at very high p: QuantilePoint(q0_975,0.975)

Examples

d = fit(LogNormal, @qp_m(3), @qp_uu(5));
quantile.(d, [0.5, 0.975]) ≈ [3,5]

DistributionFits.fit_mode_flat — Method

fit_mode_flat(::Type{LogitNormal}, mode::T; peakedness = 1.0)

Find the maximum-spread logitnormal distribution that has a single mode at given location.

More peaked distributions with given single mode can be optained by increasing argument peakedness. They will have a spread by originally inferred σ² devidied by peakedness.

Examples

m = 0.6
d = fit_mode_flat(LogitNormal, m; peakedness = 1.5)
mode(d) ≈ m

DistributionFits.meanFunOfProb — Method

compute mean over a function(uniformly distributed probabilities)

used to estimate moments of logitnorm

DistributionFits.ofLogitNormalModeUpper — Method

objective function used by matchModeUpper(LogitNormal,...)

StatsBase.fit — Method

fit(D, mean, σstar)

Fit a statistical distribution of type D to mean and multiplicative standard deviation.

Arguments

D: The type of distribution to fit

mean: The moments of the distribution

σstar::AbstractΣstar: The multiplicative standard deviation

See also σstar, AbstractΣstar.

Examples

d = fit(LogNormal, 2, Σstar(1.1));
(mean(d), σstar(d)) == (2, 1.1)